requirements.txt
aiofiles==24.1.0
aiogram==3.17.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
alembic==1.13.3
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
asteroid-filterbanks==0.4.0
async-timeout==4.0.3
attrs==24.2.0
audioread==3.0.1
av==12.3.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.8.2
contourpy==1.3.0
ctranslate2==4.5.0
cycler==0.12.1
decorator==4.4.2
docopt==0.6.2
einops==0.8.0
enum34==1.1.10
exceptiongroup==1.2.2
faster-whisper==1.0.3
filelock==3.13.1
flatbuffers==24.3.25
fonttools==4.54.1
frozenlist==1.4.1
fsspec==2024.2.0
greenlet==3.1.1
halo==0.0.31
huggingface-hub==0.26.1
humanfriendly==10.0
HyperPyYAML==1.2.2
idna==3.10
imageio==2.36.0
imageio-ffmpeg==0.5.1
iniconfig==2.0.0
Jinja2==3.1.3
joblib==1.4.2
julius==0.2.7
kiwisolver==1.4.7
lazy_loader==0.4
lightning==2.4.0
lightning-utilities==0.11.8
llvmlite==0.43.0
log-symbols==0.0.14
magic-filter==1.0.12
Mako==1.3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
MouseInfo==0.1.3
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
networkx==3.2.1
numba==0.60.0
numpy==1.26.3
omegaconf==2.3.0
onnxruntime==1.19.2
openwakeword==0.6.0
optuna==4.0.0
packaging==24.1
pandas==2.2.3
pillow==10.2.0
platformdirs==4.3.6
pluggy==1.5.0
pooch==1.8.2
primePy==1.3
proglog==0.1.10
propcache==0.2.0
protobuf==5.28.3
pvporcupine==1.9.5
pyannote.audio==3.3.2
pyannote.core==5.0.0
pyannote.database==5.1.0
pyannote.metrics==3.2.1
pyannote.pipeline==3.0.1
PyAudio==0.2.14
PyAutoGUI==0.9.54
pycparser==2.22
pydantic==2.10.5
pydantic_core==2.27.2
PyGetWindow==0.0.9
Pygments==2.18.0
PyMsgBox==1.0.9
pyparsing==3.2.0
pyperclip==1.9.0
pyreadline3==3.5.4
PyRect==0.2.0
PyScreeze==1.0.1
pytest==8.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytorch-lightning==2.4.0
pytorch-metric-learning==2.6.0
pytweening==1.2.0
pytz==2024.2
PyYAML==6.0.2
RapidFuzz==3.11.0
RealtimeSTT==0.3.5
requests==2.32.3
rich==13.9.3
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.12
schedule==1.2.2
scikit-learn==1.5.2
scipy==1.14.1
semver==3.0.2
sentencepiece==0.2.0
shellingham==1.5.4
six==1.16.0
sortedcontainers==2.4.0
soxr==0.5.0.post1
speechbrain==1.0.1
spinners==0.0.24
SQLAlchemy==2.0.36
srt==3.5.3
sympy==1.13.1
tabulate==0.9.0
tensorboardX==2.6.2.2
termcolor==2.5.0
threadpoolctl==3.5.0
tokenizers==0.20.1
tomli==2.2.1
torch==2.5.0+cu124
torch-audiomentations==0.11.1
torch_pitch_shift==1.2.5
torchaudio==2.5.0+cu124
torchmetrics==1.5.0
torchvision==0.20.0+cu124
tqdm==4.66.5
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
webrtcvad-wheels==2.0.14
websocket-client==1.8.0
websockets==12.0
yarl==1.16.0
